moves home -> pages
