+++ src/components/Body.tsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from 'react-router-dom';
+import Home from '../pages/Home';
 
-import Home from './Home';
 import DAOs from './DAOs/index';
 
 function Body() {
